requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
openai==1.51.0
twilio==9.3.0
